Split validation related tests into subfiles for better organization.
